+++ R/global_vars.R
@@ -2,8 +2,18 @@
 
 utils::globalVariables(
   names = c(
-    ".", "Group...1", "Group...2", "bf10", "ci.lower", "ci.upper",
-    "group1", "group2", "log_e_bf10", "p.crit", "p.value", "psihat"
+    ".",
+    "Group...1",
+    "Group...2",
+    "bf10",
+    "ci.lower",
+    "ci.upper",
+    "group1",
+    "group2",
+    "log_e_bf10",
+    "p.crit",
+    "p.value",
+    "psihat"
   ),
   package = "pairwiseComparisons",
   add = FALSE