+++ NEWS.md
@@ -2,6 +2,9 @@
 
   - To be consistent with the rest of the `ggstatsverse`, the Bayes Factor
     results are now always shown in favor of null over alternative (`BF01`).
+    
+  - `pairwise_comparisons` function gets `subject.id` argument relevant for
+    repeated measures design.
 
 # pairwiseComparisons 3.0.0