+++ NEWS.md
@@ -1,4 +1,4 @@
-# pairwiseComparisons 3.0.0.9000
+# pairwiseComparisons 3.1.0
 
   - To be consistent with the rest of the `ggstatsverse`, the Bayes Factor
     results are now always shown in favor of null over alternative (`BF01`).